requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-inference==1.0.0b9
azure-core==1.35.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
Flask==3.1.1
Flask-Session==0.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
msgspec==0.19.0
openai==1.97.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn